extract terminal helpers into function
