+++ src/ui/mod.rs
@@ -3,5 +3,6 @@ pub mod chat_box;
 pub mod chat_selector;
 pub mod help_box;
 pub mod input_box;
+pub mod terminal_helpers;
 pub mod title_bar;
 pub mod users;
